requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
selenium==4.19.0
yagmail==0.15.293
keyring==25.2.0 